+++ projects/pip-webui2-behaviors/src/lib/selectable/index.ts
@@ -1,0 +1,2 @@
+export * from './selectable.component';
+export { PipSelectableModule } from './selectable.module';