+++ projects/pip-webui2-behaviors/src/lib/selectable/selectable.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { PipSelectableComponent } from './selectable.component';
+import { PipSelectableDirective } from './selectable.directive';
+
+@NgModule({
+    declarations: [
+        PipSelectableComponent,
+        PipSelectableDirective
+    ],
+    imports: [
+        CommonModule
+    ],
+    exports: [
+        PipSelectableComponent,
+        PipSelectableDirective
+    ],
+    providers: [],
+})
+export class PipSelectableModule { }