+++ projects/pip-webui2-behaviors/src/public_api.ts
@@ -2,8 +2,9 @@
  * Public API Surface of pip-webui2-behaviors
  */
 
-export * from './lib/hotkeys/index';
-export * from './lib/selected/index';
-export * from './lib/focused/index';
 export * from './lib/draggable/index';
+export * from './lib/focused/index';
+export * from './lib/hotkeys/index';
 export * from './lib/infinite-scroll/index';
+export * from './lib/selectable/index';
+export * from './lib/selected/index';