+++ src/app/app-routing.module.ts
@@ -5,17 +5,17 @@ import { DraggableExampleComponent } from './draggable-example/draggable-example
 import { FocusedExampleComponent } from './focused-example/focused-example.component';
 import { InfiniteScrollExampleComponent } from './infinite-scroll-example/infinite-scroll-example.component';
 import { InfiniteScrollPageExampleComponent } from './infinite-scroll-page-example/infinite-scroll-page-example.component';
-import { SelectedExampleComponent } from './selected-example/selected-example.component';
+import { SelectableExampleComponent } from './selectable-example/selectable-example.component';
 import { HotkeysExampleComponent } from './hotkeys-example/hotkeys-example.component';
 
 const appRoutes: Routes = [
-    { path: 'selected', component: SelectedExampleComponent },
+    { path: 'selectable', component: SelectableExampleComponent },
     { path: 'focused', component: FocusedExampleComponent },
     { path: 'infinite_scroll', component: InfiniteScrollExampleComponent },
     { path: 'infinite_page', component: InfiniteScrollPageExampleComponent },
     { path: 'draggable', component: DraggableExampleComponent },
     { path: 'hotkeys', component: HotkeysExampleComponent },
-    { path: '**', redirectTo: 'selected' },
+    { path: '**', redirectTo: 'selectable' },
 ];
 
 @NgModule({