+++ src/app/app.component.spec.ts
@@ -13,7 +13,7 @@ import { ExamplesListModule } from './examples-list/examples-list.module';
 import { FocusedExampleModule } from './focused-example/focused-example.module';
 import { InfiniteScrollExampleModule } from './infinite-scroll-example/infinite-scroll-example.module';
 import { InfiniteScrollPageExampleModule } from './infinite-scroll-page-example/infinite-scroll-page-example.module';
-import { SelectedExampleModule } from './selected-example/selected-example.module';
+import { SelectableExampleModule } from './selectable-example/selectable-example.module';
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
@@ -38,7 +38,7 @@ describe('AppComponent', () => {
 
         AppRoutingModule,
         ExamplesListModule,
-        SelectedExampleModule,
+        SelectableExampleModule,
         FocusedExampleModule,
         InfiniteScrollExampleModule,
         InfiniteScrollPageExampleModule,